+++ CHANGELOG.md
@@ -8,5 +8,9 @@ In effect it is a complete rewrite of the client, with new features.
 ## 2.0.0 (2025-01-03)
 
 Version 2.0.0.alpha.2 is a bugfix release for the alpha version of the client.
-We added debugging to the api_request function.
-We also added extra information for the HTTPError exception if it was raised.
+Added debugging to the api_request function.
+Added extra information for the HTTPError exception if it was raised.
+Added access logging to the core.
+Added headers argument to the api_request and request functions to override/append headers.
+Added ETag property to the response object.
+Added etag argument to the client.request function.