+++ CHANGELOG.md
@@ -7,7 +7,7 @@ In effect it is a complete rewrite of the client, with new features.
 
 ## 2.0.0 (2025-01-03)
 
-Version 2.0.0.alpha.2 is a bugfix release for the alpha version of the client.
+Version 2.0.0.alpha.5 is a bugfix release for the alpha version of the client.
 Added debugging to the api_request function.
 Added extra information for the HTTPError exception if it was raised.
 Added access logging to the core.